Move meshnet related tests into proper file
